requirements.txt
boto3==1.34.73
botocore==1.34.73
numpy==1.26.4
pandas==2.2.1
progressbar==2.5
pymssql==2.2.11
pyodbc==5.1.0
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
s3transfer==0.10.1
SQLAlchemy==2.0.29
sendgrid==6.11.0
sentry-sdk==2.0.1